Remove `Context` from `ByteCompiler`
