+++ core/engine/src/bytecompiler/class.rs
@@ -54,7 +54,7 @@ impl ByteCompiler<'_> {
             self.json_parse,
             self.variable_environment.clone(),
             self.lexical_environment.clone(),
-            self.context,
+            self.interner,
         );
 
         compiler.code_block_flags |= CodeBlockFlags::IS_CLASS_CONSTRUCTOR;
@@ -287,7 +287,7 @@ impl ByteCompiler<'_> {
                         self.json_parse,
                         self.variable_environment.clone(),
                         self.lexical_environment.clone(),
-                        self.context,
+                        self.interner,
                     );
 
                     // Function environment
@@ -315,7 +315,7 @@ impl ByteCompiler<'_> {
                         self.json_parse,
                         self.variable_environment.clone(),
                         self.lexical_environment.clone(),
-                        self.context,
+                        self.interner,
                     );
                     let _ = field_compiler.push_compile_environment(true);
                     if let Some(node) = field {
@@ -353,7 +353,7 @@ impl ByteCompiler<'_> {
                         self.json_parse,
                         self.variable_environment.clone(),
                         self.lexical_environment.clone(),
-                        self.context,
+                        self.interner,
                     );
                     let _ = field_compiler.push_compile_environment(true);
                     if let Some(node) = field {
@@ -387,7 +387,7 @@ impl ByteCompiler<'_> {
                         false,
                         self.variable_environment.clone(),
                         self.lexical_environment.clone(),
-                        self.context,
+                        self.interner,
                     );
                     let _ = compiler.push_compile_environment(true);
