+++ core/engine/src/bytecompiler/mod.rs
@@ -21,7 +21,7 @@ use crate::{
         BindingOpcode, CodeBlock, CodeBlockFlags, Constant, GeneratorResumeKind, Handler,
         InlineCache, Opcode, VaryingOperandKind,
     },
-    Context, JsBigInt, JsString,
+    JsBigInt, JsString,
 };
 use boa_ast::{
     declaration::{Binding, LexicalDeclaration, VarDeclaration},
@@ -44,7 +44,7 @@ use rustc_hash::FxHashMap;
 use thin_vec::ThinVec;
 
 pub(crate) use declarations::{
-    eval_declaration_instantiation_context, global_declaration_instantiation_annex_b,
+    eval_declaration_instantiation_context, global_declaration_instantiation_context,
 };
 pub(crate) use function::FunctionCompiler;
 pub(crate) use jump_control::JumpControlInfo;
@@ -280,7 +280,7 @@ pub struct ByteCompiler<'ctx> {
     /// The current lexical environment.
     pub(crate) lexical_environment: Rc<CompileTimeEnvironment>,
 
-    current_open_environments_count: u32,
+    pub(crate) current_open_environments_count: u32,
     current_stack_value_count: u32,
     pub(crate) code_block_flags: CodeBlockFlags,
     handlers: ThinVec<Handler>,
@@ -296,8 +296,7 @@ pub struct ByteCompiler<'ctx> {
     pub(crate) async_handler: Option<u32>,
     json_parse: bool,
 
-    // TODO: remove when we separate scripts from the context
-    pub(crate) context: &'ctx mut Context,
+    pub(crate) interner: &'ctx mut Interner,
 
     #[cfg(feature = "annex-b")]
     pub(crate) annex_b_function_names: Vec<Identifier>,
@@ -316,8 +315,7 @@ impl<'ctx> ByteCompiler<'ctx> {
         json_parse: bool,
         variable_environment: Rc<CompileTimeEnvironment>,
         lexical_environment: Rc<CompileTimeEnvironment>,
-        // TODO: remove when we separate scripts from the context
-        context: &'ctx mut Context,
+        interner: &'ctx mut Interner,
     ) -> ByteCompiler<'ctx> {
         let mut code_block_flags = CodeBlockFlags::empty();
         code_block_flags.set(CodeBlockFlags::STRICT, strict);
@@ -346,7 +344,7 @@ impl<'ctx> ByteCompiler<'ctx> {
             json_parse,
             variable_environment,
             lexical_environment,
-            context,
+            interner,
 
             #[cfg(feature = "annex-b")]
             annex_b_function_names: Vec::new(),
@@ -370,7 +368,7 @@ impl<'ctx> ByteCompiler<'ctx> {
     }
 
     pub(crate) fn interner(&self) -> &Interner {
-        self.context.interner()
+        self.interner
     }
 
     fn get_or_insert_literal(&mut self, literal: Literal) -> u32 {
@@ -1320,7 +1318,7 @@ impl<'ctx> ByteCompiler<'ctx> {
                 body,
                 self.variable_environment.clone(),
                 self.lexical_environment.clone(),
-                self.context,
+                self.interner,
             );
 
         self.push_function_to_constants(code)
@@ -1395,7 +1393,7 @@ impl<'ctx> ByteCompiler<'ctx> {
                 body,
                 self.variable_environment.clone(),
                 self.lexical_environment.clone(),
-                self.context,
+                self.interner,
             );
 
         let index = self.push_function_to_constants(code);
@@ -1442,7 +1440,7 @@ impl<'ctx> ByteCompiler<'ctx> {
                 body,
                 self.variable_environment.clone(),
                 self.lexical_environment.clone(),
-                self.context,
+                self.interner,
             );
 
         let index = self.push_function_to_constants(code);
